import

+++ app/Http/Controllers/BooksController.php
@@ -3,19 +3,25 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreBookRequest;
+use App\Http\Requests\StoreImportRequest;
 use App\Models\Book;
+use App\Models\Import;
 use App\Models\Loaisach;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BooksController extends Controller
 {
     protected $book;
     protected $loaisach;
+    protected $import;
 
-    public function __construct(Book $book, Loaisach $loaisach)
+    public function __construct(Book $book, Loaisach $loaisach, Import $import)
     {
         $this->book = $book;
         $this->loaisach = $loaisach;
+        $this->import = $import;
         $this->middleware('auth')->except(['index', 'show']);
     }
 
@@ -42,7 +48,10 @@ class BooksController extends Controller
      */
     public function create()
     {
-        return view('books.create');
+        if (Gate::any(['admin', 'staff'], Auth::user())) {
+            return view('books.create');
+        }
+        return redirect()->route('home');
     }
 
     /**
@@ -53,6 +62,9 @@ class BooksController extends Controller
      */
     public function store(StoreBookRequest $request)
     {
+        if (Gate::none(['admin', 'staff'], Auth::user())) {
+            return redirect()->route('books.index');
+        }
         $this->book->saveBook($request);
 
         flash('add success')->success();
@@ -81,8 +93,11 @@ class BooksController extends Controller
      */
     public function edit($id)
     {
-        $book = $this->book->find($id);
-        return view('books.edit')->with('book', $book);
+        if (Gate::any(['admin', 'staff'], Auth::user())) {
+            $book = $this->book->find($id);
+            return view('books.edit')->with('book', $book);
+        }
+        return redirect()->route('books.show', $id);
     }
 
     /**
@@ -91,13 +106,16 @@ class BooksController extends Controller
      * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request)
+    public function update(StoreBookRequest $request)
     {
+        if (Gate::none(['admin', 'staff'], Auth::user())) {
+            return redirect()->route('books.show', $request->id);
+        }
+
         $this->book->updateBook($request);
 
         flash('update success')->success();
-
-        return redirect()->route('books.index');
+        return redirect()->route('books.show', $request->id);
     }
 
     /**
@@ -108,10 +126,31 @@ class BooksController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::none(['admin', 'staff'], Auth::user())) {
+            return redirect()->route('books.show', $id);
+        }
+
         $this->book->find($id)->delete();
 
         flash('delete success')->error();
-
         return redirect()->route('books.index');
+    }
+
+    public function storeImportRequest(StoreImportRequest $request, $book_id)
+    {
+        if (!Gate::any(['admin', 'staff', 'warehouseman'], Auth::user())) {
+            return redirect()->route('books.show', $book_id);
+        }
+
+        $data = $request->all();
+        $data['book_id']=$book_id;
+        $data['user_id']=Auth::id();
+        $data['status'] = 0;
+        $data['accepted_id'] = 1;
+
+        $this->import->create($data);
+
+        flash('Send import request succeeded');
+        return redirect()->route('books.show', $book_id);
     }
 }
